Add RPC method validate
