+++ src/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import cors from 'cors';
-import rpc from './rpc';
+import api from './api';
 
 const app = express();
 const PORT = process.env.PORT || 3000;
@@ -8,6 +8,6 @@ const PORT = process.env.PORT || 3000;
 app.use(express.json({ limit: '8mb' }));
 app.use(express.urlencoded({ limit: '8mb', extended: false }));
 app.use(cors({ maxAge: 86400 }));
-app.use('/', rpc);
+app.use('/', api);
 
 app.listen(PORT, () => console.log(`Listening at http://localhost:${PORT}`));
